+++ components/organisms/ui/index.tsx
@@ -3,9 +3,9 @@ import { useCallback, useState } from 'react'
 import { useIdolData } from 'hooks/useIdolData'
 import { useKeepId } from 'hooks/useKeepId'
 
-import MoveTopButton from './button/move-top'
-import Search from './search'
-import SearchResults from './search-results'
+import MoveTopButton from '../button/move-top'
+import ColorCards from '../color-cards'
+import Search from '../search'
 
 const UI = () => {
   const [bland, setBland] = useState('')
@@ -31,7 +31,7 @@ const UI = () => {
         onChangeName={handleChangeName}
         onChangeSimilarColor={handleChangeSimilarColor}
       />
-      <SearchResults
+      <ColorCards
         className="mt-12"
         items={searchResults}
         keepIdList={keepIdList}
